+++ rama-net/src/lib.rs
@@ -23,6 +23,7 @@ pub mod conn;
 pub mod forwarded;
 pub mod mode;
 pub mod stream;
+pub mod test_utils;
 pub mod user;
 
 pub(crate) mod proto;
