+++ rama-net/src/test_utils/client/mod.rs
@@ -1,0 +1,2 @@
+mod mock_connector;
+pub use mock_connector::MockConnectorService;
